pyproject.toml
[tool.poetry]
name = "otk"
version = "0.1.0"
description = "Optics Tool Kit"
authors = ["Dane Austin <dane_austin@fastmail.com.au>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/draustin/otk"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.18.4"
numba = "^0.49.1"
pyqt5 = "^5.12"
pyopengl = "^3.1.5"
mathx = "^0.2"
pyyaml = "^5.3.1"
pyfftw = "^0.12.0"
opt_einsum = "^3.2.1"
pyqtgraph_extensions = "^0.4.0"
gizeh = "^0.1.11"
matplotlib = {version = "^3.2.1"}
chardet = "^3.0.4"
# Ideally these would be optional. But I'm getting weird behaviour suggestive of bugs in poetry, so want to keep it simple for now.
# Spyder is currently a pain to install. See https://github.com/python-poetry/poetry/issues/1238. Tried various workarounds
# but the nonhalting dependency resolution it was taking too long and I gave up.
#spyder-kernels = { git = "https://github.com/danieltomasz/spyder-kernels.git", branch = "poetry", optional = false}
#spyder = {version = "^4.1", optional = false}
ipython = {version = "^7.14.0", optional = false}
cairocffi = "^1.1.0"


[tool.poetry.scripts]
view-zmx = 'otk.rt2.cli:view_zmx'


[tool.poetry.dev-dependencies]
pytest = "^5.4.2"
pytest-qt = "^3.3.0"

# To generate setup.py, install dephell and then `dephell deps convert`.
[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"